add: Lualine for an epic statusline + fix: goyo
fix: As it turns out, Limelight hadn't been launching this entire
time because the function overrides the autocmd or something. I found
this out when I tried to add the code to disable lualine to the autocmd.

add: lualine, for a hipster status bar. I had to install a patched font,
but it was worth it because it looks cool I think. Idk why such a
popular statusline isn't built-in to Goyo, but I'm sure glad it wasn't
because it let me discover the fact that Limelight hadn't been launching
lol.