add: Install Python pkgs to virtual env bootstrap
The aliases are already in ~/.zshrc, I just never bothered to add a
bootstrapping step or provide instructions.

Though, this does controversially create ~/Scripts, which "should" go
into ~/Documents or ~/.config, I don't want to do that because I want to
it be a ~ folder like ~/Downloads or ~/Music. Super dumb and
opinionated, I know, but this is my dotfiles repo.