replace: vim-grammarous with ltex-ls lsp
ltex-ls is basically a LSP server for language tool.

the only downside is vim-grammarous's UI for grammar checking that would
display the offending snippet with nice colors and an explanation of
what's wrong—i'll need to create/find a telescope integration
that can do that (similar to my z= spellsuggest override) and/or find a
better way to read LSP messages (e.g. folke/trouble.nvim)

the spam from noice about ltex running every time I type a character
also needs to be fixed, along with the fact that ltex appears to be
running on every buffer type.